==> scripts/control_totals/old_scripts/06_Testing.R <==
library(tidyr)
library(data.table)
##get rid of scientific notation
options(scipen=999)
setwd('S:/Network Shares/DS Projects/Current_Projects/Projections/Reweighter/Scripts')

source('read_config_hh.R')
source('Manage_Scenario.R')
